requirements.txt
Flask==2.1.1
gunicorn==20.1.0

requests==2.27.1

google-auth==2.3.2
google-cloud-logging==3.1.1
